:truck: (backend) move users module to modeules/user
