+++ apps/backend/src/auth/commands/login/login.command.ts
@@ -1,5 +1,5 @@
 import { Command } from '@egodb/domain'
-import { User } from '../../../users/users.service'
+import { User } from '../../../modules/user/user.service'
 import { LoginDTO } from '../../dtos/login.dto'
 
 export class LoginCommand extends Command implements LoginDTO {
