+++ apps/backend/src/auth/dtos/get-me.response.dto.ts
@@ -1,4 +1,4 @@
-import { User } from '../../users/users.service'
+import { User } from '../../modules/user/user.service'
 
 export interface GetMeResponseDTO {
   me: User
