+++ apps/backend/src/auth/dtos/login.dto.ts
@@ -1,4 +1,4 @@
-import { User } from '../../users/users.service'
+import { User } from '../../modules/user/user.service'
 
 export interface LoginDTO {
   user: User
