+++ apps/backend/src/auth/queries/me/get-me.query.ts
@@ -1,5 +1,5 @@
 import { Query } from '@egodb/domain'
-import { User } from '../../../users/users.service.js'
+import { User } from '../../../modules/user/user.service.js'
 import { GetMeDTO } from '../../dtos/get-me.dto.js'
 
 export class GetMeQuery extends Query implements GetMeDTO {
