+++ apps/backend/src/modules/index.ts
@@ -1,3 +1,4 @@
 import { TableModule } from './table/table.module.js'
+import { UserModule } from './user/user.module.js'
 
-export const modules = [TableModule]
+export const modules = [TableModule, UserModule]
