+++ apps/backend/src/modules/user/user.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common'
 export type User = any
 
 @Injectable()
-export class UsersService {
+export class UserService {
   private readonly users = [
     {
       userId: 1,
